Use redisson bucket cache for all caches
The map cache is easy to structure the relevant caches together but needs better eviction policy to get rid of the caches when reach threshold. Before having that, we should switch to bucket cache across all categories.
