+++ core/src/main/java/org/mskcc/cbio/oncokb/cache/CacheNameResolver.java
@@ -3,7 +3,7 @@ package org.mskcc.cbio.oncokb.cache;
 import org.mskcc.cbio.oncokb.util.PropertiesUtils;
 import org.springframework.stereotype.Component;
 
-import static org.mskcc.cbio.oncokb.cache.Constants.DELIMITER;
+import static org.mskcc.cbio.oncokb.cache.Constants.REDIS_KEY_SEPARATOR;
 
 @Component
 public class CacheNameResolver {
@@ -13,6 +13,6 @@ public class CacheNameResolver {
     }
 
     public String getCacheName(String cacheKey) {
-        return this.appName + DELIMITER + cacheKey;
+        return this.appName + REDIS_KEY_SEPARATOR + cacheKey;
     }
 }
