+++ core/src/main/java/org/mskcc/cbio/oncokb/cache/Constants.java
@@ -1,5 +1,5 @@
 package org.mskcc.cbio.oncokb.cache;
 
 public final class Constants {
-    public static final String DELIMITER = ":";
+    public static final String REDIS_KEY_SEPARATOR = ":";
 }
